% Auto-generated: do not edit by hand
\name{ppdv}

\alias{ppdv}

\title{Ppdv component}

\description{
ExampleComponent is an example component. It takes a property, `label`, and displays it. It renders an input with the property `value` which is editable by the user.
}

\usage{
ppdv(id=NULL, sensorData=NULL)
}

\arguments{
\item{id}{Character. }

\item{sensorData}{List of lists containing elements 'id', 'name', 'value'.
those elements have the following types:
  - id (numeric; required)
  - name (character; required)
  - value (numeric; required)s. }
}

\value{named list of JSON elements corresponding to React.js properties and their values}

